+++ src/app/(main)/design-system/button-demo.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import React from 'react'
-import Button from '@/components/Button'
-import ButtonGroup from '@/components/ButtonGroup'
+import Button from '@/components/ui/Button'
+import ButtonGroup from '@/components/ui/ButtonGroup'
 import { StarIcon, HeartIcon, PlusIcon } from '@/components/icons'
 import { DESIGN_SYSTEM_CONSTANTS, getSpacingClass } from './constants'
