+++ src/app/(main)/design-system/misc-demo.tsx
@@ -1,15 +1,15 @@
 'use client'
 
 import React, { useState } from 'react'
-import Tab from '@/components/Tab'
-import TabItem from '@/components/TabItem'
-import Slider from '@/components/Slider'
-import Tag from '@/components/Tag'
-import HelpText from '@/components/HelpText'
-import StatusLight from '@/components/StatusLight'
-import AlertBanner from '@/components/AlertBanner'
-import AlertDialog from '@/components/AlertDialog'
-import Badge from '@/components/Badge'
+import Tab from '@/components/ui/Tab'
+import TabItem from '@/components/ui/TabItem'
+import Slider from '@/components/ui/Slider'
+import Tag from '@/components/ui/Tag'
+import HelpText from '@/components/ui/HelpText'
+import StatusLight from '@/components/ui/StatusLight'
+import AlertBanner from '@/components/ui/AlertBanner'
+import AlertDialog from '@/components/ui/AlertDialog'
+import Badge from '@/components/ui/Badge'
 import { StarIcon } from '@/components/icons'
 import { DESIGN_SYSTEM_CONSTANTS, getSpacingClass } from './constants'
