+++ src/components/Editor/VideoPlayer/index.tsx
