+++ src/components/Header.tsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
-import Button from '@/components/Button'
+import Button from '@/components/ui/Button'
 
 interface HeaderProps {
   isVisible?: boolean
