+++ src/components/LandingPage/OpenLibrarySection.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Button from '@/components/Button'
+import Button from '@/components/ui/Button'
 
 interface OpenLibrarySectionProps {
   libraryRef: React.RefObject<HTMLElement | null>
