+++ src/components/UploadModal.tsx
@@ -4,10 +4,10 @@ import React, { useState, useCallback, useMemo, useRef } from 'react'
 
 import { cn, SIZE_CLASSES } from '@/lib/utils'
 
-import Modal, { type ModalProps } from './Modal'
-import Button from './Button'
-import ButtonGroup from './ButtonGroup'
-import ProgressCircle from './ProgressCircle'
+import Modal, { type ModalProps } from './ui/Modal'
+import Button from './ui/Button'
+import ButtonGroup from './ui/ButtonGroup'
+import ProgressCircle from './ui/ProgressCircle'
 import FileUploadTab from './UploadModal/FileUploadTab'
 import UrlImportTab from './UploadModal/UrlImportTab'
 import TranscriptionSettings from './UploadModal/TranscriptionSettings'
