+++ src/components/UploadModal/FileUploadTab.tsx
@@ -1,8 +1,8 @@
 'use client'
 
 import React from 'react'
-import Button from '@/components/Button'
-import Badge from '@/components/Badge'
+import Button from '@/components/ui/Button'
+import Badge from '@/components/ui/Badge'
 
 interface FileUploadTabProps {
   dragActive: boolean
